Reaction norm plots for entire simulation
This version allows the import of the entire experiment in one script.
Several steps are time consuming (GRN import, GRN develop and problem import)
We annotate the replicate of origin for each test simulation based on directory structure. This could be replaced with GRN filename for robustness.
Currently only plots a main overview, to check that all data are in the right place for generating the subplots in the paper.